Be explicit that Fos32R is an unsigned value.

+++ src/main/java/com/sibilantsolutions/grison/driver/foscam/mapper/DtoToEntity.java
@@ -177,10 +177,10 @@ public final class DtoToEntity {
         final InetAddress dns;
 
         try {
-            address = new InetSocketAddress(InetAddress.getByAddress(Ints.toByteArray(dto.ip().value())), dto.cameraPort().value());
-            mask = InetAddress.getByAddress(Ints.toByteArray(dto.mask().value()));
-            gateway = InetAddress.getByAddress(Ints.toByteArray(dto.gateway().value()));
-            dns = InetAddress.getByAddress(Ints.toByteArray(dto.dns().value()));
+            address = new InetSocketAddress(InetAddress.getByAddress(Ints.toByteArray(dto.ip().value().intValue())), dto.cameraPort().value());
+            mask = InetAddress.getByAddress(Ints.toByteArray(dto.mask().value().intValue()));
+            gateway = InetAddress.getByAddress(Ints.toByteArray(dto.gateway().value().intValue()));
+            dns = InetAddress.getByAddress(Ints.toByteArray(dto.dns().value().intValue()));
         } catch (UnknownHostException e) {
             throw new UnsupportedOperationException("TODO (CSB)", e);
         }
